backend/requirements.txt
Flask==2.2.5
mysql-connector-python==8.1.0

